Trait trillium_askama::AskamaConnExt

pub trait AskamaConnExt {
    fn render(self, template: impl Template) -> Self;
}

extends trillium conns with the ability to render askama templates

Required Methods

fn render(self, template: impl Template) -> Self

renders an askama template, halting the conn and setting a 200 status code. also sets the mime type based on the template extension